requirements.txt
fastapi
uvicorn
pymupdf
langchain
langchain_community
langchain-huggingface
chromadb
ollama
python-dotenv

